requirements.txt
fastapi
uvicorn
pymongo
motor
faiss-cpu
sentence-transformers
pandas
numpy
python-multipart
PyPDF2
beautifulsoup4
google-generativeai
python-dotenv
scikit-learn
openpyxl
pydantic
# LangChain dependencies
langchain
langchain-google-genai
langchain-community